knowledge-base/weaviate-agent/requirements.txt
fastapi
uvicorn
weaviate-client
pypdf